pyproject.toml
[tool.poetry]
name = "searchright-technical-assignment"
version = "0.1.0"
description = ""
authors = ["Nuung <qlgks1@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
psycopg2-binary = "^2.9.10"
openai = "^1.75.0"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
